fix: Create Date Funcations to calculation age and valid birth
