--- src/utils/birthFormatter.tsx
@@ -1,12 +1,0 @@
-export const birthFomatter = (input: string): string => {
-  // 입력된 8자리 숫자를 문자열로 변환하여 처리합니다.
-  const dateStr = input.toString();
-
-  // 년, 월, 일 부분을 추출합니다.
-  const year = dateStr.slice(0, 4);
-  const month = dateStr.slice(4, 6);
-  const day = dateStr.slice(6, 8);
-
-  // 포맷된 문자열을 반환합니다.
-  return `${year}년 ${month}월 ${day}일`;
-};
